fix: sanitize NNBF characters in uploaded files
MacOS 14.0+ uses a narrow no-break space (U+202F) in the timestamp of
automatically named screenshots (12-hour clock only). This is not a
valid filename character for storage, and screenshots are probably a
common-ish use case, so treat this as a special case and replace with
a normal space.

See: https://forum.obsidian.md/t/unable-to-attach-screenshot-images-in-macos-14-0-when-wikilinks-turned-off-urlencode-u-202f-char/68410/6